requirements.txt
google-genai
fastapi
uvicorn
python-dotenv
langgraph
langchain-core
pydub
demucs
soundfile
